+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Messages pages - for displaying user prompt - responses and files to the user
 - Files and messages are highlighted visually when selected as reference
 - Files and messages can be de-selected from the display area over the prompt
+- Categories are automatically assigned a random colour with an option for an LLM assigned colour
 
 ### Changed